pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "climatic-water-balance"
version = "0.1.0"
description = "Calculate climatic water balance using the Thornthwaite method"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "xarray>=0.20.0",
    "rioxarray>=0.9.0",
    "dask>=2022.1.0",
    "zarr>=2.10.0",
    "rasterio>=1.2.0",
    "pyyaml>=6.0",
    "openpyxl>=3.0.0", # For Excel file support
    "bottleneck>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.1.0",
    "isort>=5.10.0",
    "mypy>=0.931",
    "flake8>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "rasterio.*",
    "rioxarray.*",
    "xarray.*",
    "dask.*",
    "zarr.*",
]
ignore_missing_imports = true
